Исправил замечания
